requirements.txt
annotated-types==0.7.0
anthropic==0.78.0
bm25s==0.2.14
pystemmer==3.0.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
docstring-parser==0.17.0
filelock==3.20.3
fsspec==2026.2.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.4.0
hyperframe==6.1.0
idna==3.11
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
markdownify==1.2.2
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
openai==2.17.0
packaging==26.0
portalocker==3.2.0
protobuf==6.33.5
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
python-dotenv==1.2.1
pyyaml==6.0.3
qdrant-client==1.16.2
rapidfuzz==3.14.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
